refactor(core): extract EVM tx execution + read-only validate from block_executor
block_executor.rs was 2532 LOC and the navigation around the apply
hot path was rough — every time you wanted to read what Pass 2
actually does, you had to scroll past 250 lines of EVM tx decoding
and another 260 lines of read-only validation that the BFT engine
calls under a shared lock.

Pulled both into their own files under block_executor/:

- block_executor/evm.rs (269 LOC): the per-tx EVM execution that
  runs inside Pass 2 whenever a tx's data field starts with 'EVM:'.
  Re-decodes the alloy RLP envelope, runs revm, writes back balance
  / storage / code deltas. Failure marks the receipt status=0x0
  rather than rolling back the whole block.

- block_executor/validate.rs (276 LOC): the read-lock-only
  validate_block path the BFT engine + RPC use to reject obviously-
  broken blocks before contending for the chain write lock. Mirrors
  Pass 1 of add_block_impl which is still the single source of truth.

block_executor.rs drops 2532 → 2018 LOC. add_block / add_block_impl
/ apply_block_pass2 stay in the parent file because they're the
real apply hot path; splitting them further would just hide the
two-pass shape behind another layer of indirection.

cargo clippy --workspace --tests -- -D warnings clean.
cargo test --workspace passes.
