+++ crates/sentrix-core/src/block_executor.rs
@@ -1,5 +1,11 @@
 // block_executor.rs - Sentrix — Block validation and commit (two-pass)
 
+// Per-tx EVM execution and the read-only validate path live in their
+// own files so this one can stay focused on the block-level apply
+// hot path.
+mod evm;
+mod validate;
+
 use crate::authority::AuthorityManager;
 use crate::blockchain::{
     Blockchain, CHAIN_WINDOW_SIZE, is_spendable_sentrix_address, is_valid_sentrix_address,
@@ -102,269 +108,6 @@ fn shadow_observe_parallel_batching(block: &Block) {
 }
 
 impl Blockchain {
-    /// P1 (write-lock scope split): pure read-only validation of a block
-    /// against the current chain state. Safe to call under a shared
-    /// `RwLock::read()` guard — performs no mutations. Mirrors Pass 1 of
-    /// `add_block` so that callers can reject obviously-invalid blocks
-    /// (wrong height, bad signature, insufficient balance, duplicate
-    /// nonce, etc.) WITHOUT needing the write lock that would starve
-    /// RPC readers for the duration of the check. On success the
-    /// caller takes the write lock and calls `add_block` for the
-    /// commit; `add_block` re-runs Pass 1 internally as the single
-    /// source of truth for safety, so this pre-flight is an
-    /// optimisation only — never a correctness substitute.
-    #[allow(clippy::cognitive_complexity)]
-    pub fn validate_block(&self, block: &Block) -> SentrixResult<()> {
-        let expected_index = self.height() + 1;
-        let expected_prev = self.latest_block()?.hash.clone();
-
-        block.validate_structure(expected_index, &expected_prev)?;
-
-        // 2026-04-26 fix: use the runtime-aware voyager_mode_for() check
-        // (ORs env-var fork-height with chain.db `voyager_activated` flag)
-        // instead of the static is_voyager_height(). Prevents the 2026-04-26
-        // mainnet stall where env var defaulted to u64::MAX, the static
-        // check returned false, and Pioneer auth rejected legit skip-round
-        // Voyager blocks. See `incidents/2026-04-26-voyager-fork-height-env-bug.md`.
-        if !self.voyager_mode_for(expected_index)
-            && !self
-                .authority
-                .is_authorized(&block.validator, expected_index)?
-        {
-            return Err(SentrixError::UnauthorizedValidator(format!(
-                "validator {} not authorized for block {}",
-                block.validator, expected_index
-            )));
-        }
-
-        let prev_timestamp = self.latest_block()?.timestamp;
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-        if block.timestamp < prev_timestamp {
-            return Err(SentrixError::InvalidBlock(
-                "block timestamp is before previous block".to_string(),
-            ));
-        }
-        if block.timestamp > now + 15 {
-            return Err(SentrixError::InvalidBlock(
-                "block timestamp too far in the future".to_string(),
-            ));
-        }
-
-        // Phase D Q4 (required-presence): post-fork at epoch boundaries, if
-        // local LivenessTracker shows downtime evidence, the block MUST
-        // contain a JailEvidenceBundle system tx. Otherwise the proposer
-        // omitted a required system op — reject the block.
-        //
-        // Symmetry note: if local_evidence is empty, dispatch (block_executor
-        // apply path) already rejects blocks that include a JailEvidenceBundle
-        // whose claimed evidence diverges from local recompute. So we only
-        // need to enforce presence here, not absence.
-        //
-        // Pre-fork (default JAIL_CONSENSUS_HEIGHT=u64::MAX): this branch is
-        // unreachable, so behavior is unchanged on default builds.
-        if Self::is_jail_consensus_height(expected_index)
-            && sentrix_staking::epoch::EpochManager::is_epoch_boundary(expected_index)
-        {
-            let active_set = self.stake_registry.active_set.clone();
-            // 2026-04-29: pass current_height so the deterministic
-            // is_downtime_at check has the window cutoff it needs.
-            let local_evidence = self
-                .slashing
-                .compute_jail_evidence(&active_set, expected_index);
-            if !local_evidence.is_empty()
-                && !block.transactions.iter().any(|tx| tx.is_system_tx())
-            {
-                return Err(SentrixError::InvalidBlock(format!(
-                    "boundary block {} missing required JailEvidenceBundle \
-                     (local evidence: {} entries — proposer omitted required system op)",
-                    expected_index,
-                    local_evidence.len()
-                )));
-            }
-        }
-
-        let reward = self.get_block_reward();
-        let coinbase = block
-            .coinbase()
-            .ok_or_else(|| SentrixError::InvalidBlock("missing coinbase".to_string()))?;
-        if coinbase.amount != reward {
-            return Err(SentrixError::InvalidBlock(format!(
-                "coinbase amount {} must equal block reward {}",
-                coinbase.amount, reward
-            )));
-        }
-        if coinbase.to_address != block.validator {
-            return Err(SentrixError::InvalidBlock(format!(
-                "coinbase recipient {} must equal block validator {}",
-                coinbase.to_address, block.validator
-            )));
-        }
-
-        let mut working_balances: HashMap<String, u64> = HashMap::new();
-        let mut working_nonces: HashMap<String, u64> = HashMap::new();
-        let mut seen_sender_nonce: HashSet<(String, u64)> = HashSet::new();
-
-        for tx in block.transactions.iter().skip(1) {
-            // Phase D: system-emitted txs (JailEvidenceBundle from PROTOCOL_TREASURY)
-            // skip standard nonce/balance validation. Auth is consensus-driven:
-            // verified at apply via recompute-and-compare in block_executor.
-            if tx.is_system_tx() {
-                continue;
-            }
-
-            if !seen_sender_nonce.insert((tx.from_address.clone(), tx.nonce)) {
-                return Err(SentrixError::InvalidBlock(format!(
-                    "duplicate (sender, nonce) pair for {} nonce {} in block",
-                    tx.from_address, tx.nonce
-                )));
-            }
-
-            let balance = working_balances
-                .get(&tx.from_address)
-                .copied()
-                .unwrap_or_else(|| self.accounts.get_balance(&tx.from_address));
-
-            let nonce = working_nonces
-                .get(&tx.from_address)
-                .copied()
-                .unwrap_or_else(|| self.accounts.get_nonce(&tx.from_address));
-
-            tx.validate(nonce, self.chain_id)?;
-
-            let needed = tx.amount.checked_add(tx.fee).ok_or_else(|| {
-                SentrixError::InvalidTransaction("amount + fee overflow".to_string())
-            })?;
-            if balance < needed {
-                return Err(SentrixError::InsufficientBalance {
-                    have: balance,
-                    need: needed,
-                });
-            }
-
-            if let Some(token_op) = TokenOp::decode(&tx.data) {
-                match &token_op {
-                    TokenOp::Transfer {
-                        contract,
-                        to,
-                        amount,
-                    } => {
-                        if !self.contracts.exists(contract) {
-                            return Err(SentrixError::InvalidTransaction(format!(
-                                "token contract {} not found",
-                                contract
-                            )));
-                        }
-                        if !is_spendable_sentrix_address(to) {
-                            return Err(SentrixError::InvalidTransaction(format!(
-                                "invalid token transfer target address: '{}' \
-                                 (zero address rejected — use Burn op to destroy tokens)",
-                                to
-                            )));
-                        }
-                        let token_bal =
-                            self.contracts.get_token_balance(contract, &tx.from_address);
-                        if token_bal < *amount {
-                            return Err(SentrixError::InsufficientBalance {
-                                have: token_bal,
-                                need: *amount,
-                            });
-                        }
-                    }
-                    TokenOp::Burn { contract, amount } => {
-                        if !self.contracts.exists(contract) {
-                            return Err(SentrixError::InvalidTransaction(format!(
-                                "token contract {} not found",
-                                contract
-                            )));
-                        }
-                        let token_bal =
-                            self.contracts.get_token_balance(contract, &tx.from_address);
-                        if token_bal < *amount {
-                            return Err(SentrixError::InsufficientBalance {
-                                have: token_bal,
-                                need: *amount,
-                            });
-                        }
-                    }
-                    TokenOp::Mint { contract, to, .. } => {
-                        if !self.contracts.exists(contract) {
-                            return Err(SentrixError::InvalidTransaction(format!(
-                                "token contract {} not found",
-                                contract
-                            )));
-                        }
-                        if !is_spendable_sentrix_address(to) {
-                            return Err(SentrixError::InvalidTransaction(format!(
-                                "invalid token mint target address: '{}' (zero \
-                                 address rejected)",
-                                to
-                            )));
-                        }
-                    }
-                    TokenOp::Approve {
-                        contract, spender, ..
-                    } => {
-                        if !self.contracts.exists(contract) {
-                            return Err(SentrixError::InvalidTransaction(format!(
-                                "token contract {} not found",
-                                contract
-                            )));
-                        }
-                        if !is_valid_sentrix_address(spender) {
-                            return Err(SentrixError::InvalidTransaction(format!(
-                                "invalid token approve spender address: '{}'",
-                                spender
-                            )));
-                        }
-                    }
-                    TokenOp::Deploy { name, symbol, .. } => {
-                        if name.is_empty() || name.len() > 64 {
-                            return Err(SentrixError::InvalidTransaction(
-                                "token name must be 1–64 characters".to_string(),
-                            ));
-                        }
-                        if symbol.is_empty()
-                            || symbol.len() > 10
-                            || !symbol.chars().all(|c| c.is_ascii_alphanumeric())
-                        {
-                            return Err(SentrixError::InvalidTransaction(
-                                "token symbol must be 1–10 ASCII alphanumeric characters"
-                                    .to_string(),
-                            ));
-                        }
-                    }
-                    op if op.is_nft_family() => {
-                        // SRC-721 + SRC-1155 dispatch is gated by
-                        // NFT_TOKENOP_HEIGHT fork. Pre-fork: reject. Wire
-                        // format stable from this PR; storage layer +
-                        // dispatch land in the follow-up PR.
-                        if !Self::is_nft_tokenop_height(self.height() + 1) {
-                            return Err(SentrixError::InvalidTransaction(
-                                "NFT TokenOp dispatch is gated by \
-                                 NFT_TOKENOP_HEIGHT fork (currently disabled)"
-                                    .into(),
-                            ));
-                        }
-                    }
-                    _ => unreachable!("TokenOp variant handled above"),
-                }
-            }
-
-            *working_balances
-                .entry(tx.from_address.clone())
-                .or_insert(balance) -= needed;
-            *working_nonces
-                .entry(tx.from_address.clone())
-                .or_insert(nonce) += 1;
-        }
-
-        Ok(())
-    }
-
-    // ── Block application (two-pass atomic) ─────────────
     /// Admit a block produced locally. Preserves existing call sites that
     /// don't care about origin (tests, legacy integrations). For blocks
     /// arriving from peers past `STATE_ROOT_FORK_HEIGHT`, use
@@ -1649,261 +1392,6 @@ impl Blockchain {
         self.stake_registry.delegator_rewards.clear();
     }
 
-    /// Execute an EVM transaction (from eth_sendRawTransaction) within a block.
-    /// Decodes the original RLP tx from the signature field, runs it through revm,
-    /// applies state changes (contract creation, storage updates, balance transfers).
-    fn execute_evm_tx_in_block(
-        &mut self,
-        tx: &sentrix_primitives::transaction::Transaction,
-        block_height: u64,
-        block_hash_hex: &str,
-        tx_index: u32,
-    ) -> SentrixResult<()> {
-        // Parse "EVM:gas_limit:hex_data" from data field
-        let parts: Vec<&str> = tx.data.splitn(3, ':').collect();
-        if parts.len() != 3 || parts[0] != "EVM" {
-            return Ok(()); // not an EVM tx, skip
-        }
-        let gas_limit: u64 = parts[1].parse().unwrap_or(30_000_000);
-        let calldata = hex::decode(parts[2]).unwrap_or_default();
-
-        // Decode raw Ethereum tx from signature field for re-validation
-        let raw_bytes = match hex::decode(&tx.signature) {
-            Ok(b) => b,
-            Err(_) => return Ok(()), // malformed, skip silently
-        };
-
-        use alloy_consensus::TxEnvelope;
-        use alloy_consensus::transaction::SignerRecoverable;
-        use alloy_eips::eip2718::Decodable2718;
-
-        let envelope: TxEnvelope = match TxEnvelope::decode_2718(&mut raw_bytes.as_slice()) {
-            Ok(e) => e,
-            Err(_) => return Ok(()),
-        };
-        let _sender = envelope.recover_signer().ok();
-
-        // Build EVM tx
-        use alloy_primitives::{B256, U256};
-        use revm::context::TxEnv;
-        use revm::primitives::TxKind;
-        use revm::state::Bytecode;
-        use sentrix_evm::database::{SentrixEvmDb, parse_sentrix_address};
-        use sentrix_evm::executor::execute_tx_with_state;
-        use sentrix_evm::gas::INITIAL_BASE_FEE;
-        use sentrix_evm::writeback::commit_state_to_account_db;
-
-        let from_addr =
-            parse_sentrix_address(&tx.from_address).unwrap_or(alloy_primitives::Address::ZERO);
-        let to_addr_str = if tx.to_address == sentrix_primitives::transaction::TOKEN_OP_ADDRESS {
-            None
-        } else {
-            parse_sentrix_address(&tx.to_address)
-        };
-        let tx_kind = match to_addr_str {
-            Some(addr) => TxKind::Call(addr),
-            None => TxKind::Create,
-        };
-
-        // Build EVM db from CURRENT AccountDB so the EVM sees real balances,
-        // nonces, and contract code pointers — not a fresh InMemoryDB as pre-fix.
-        // from_account_db handles sentri → wei scaling for all accounts.
-        let mut evm_db = SentrixEvmDb::from_account_db(&self.accounts);
-        let sender_nonce = self.accounts.get_nonce(&tx.from_address);
-
-        // For Call txs, pre-load the target contract's bytecode + existing
-        // storage slots. Without this, revm sees an empty-code address and
-        // executes nothing useful. CREATE txs don't need pre-load — code
-        // comes from tx data and gets deployed by revm itself.
-        if let TxKind::Call(target_addr) = tx_kind {
-            let target_str = format!("0x{}", hex::encode(target_addr.as_slice()));
-            if let Some(target_account) = self.accounts.accounts.get(&target_str)
-                && target_account.code_hash != sentrix_primitives::EMPTY_CODE_HASH
-            {
-                let code_hash_hex = hex::encode(target_account.code_hash);
-                if let Some(bytecode) = self.accounts.get_contract_code(&code_hash_hex) {
-                    let code = Bytecode::new_raw(alloy_primitives::Bytes::from(bytecode.clone()));
-                    evm_db.insert_code(B256::from(target_account.code_hash), code);
-                }
-                // Pre-load storage slots whose key starts with "{target_addr}:".
-                // Sorted by slot_hex so the insertion order is deterministic
-                // across validator processes (HashMap iteration otherwise is
-                // non-deterministic; same pattern as init_trie backfill).
-                let prefix = format!("{}:", target_str);
-                let mut slots: Vec<(String, Vec<u8>)> = self
-                    .accounts
-                    .contract_storage
-                    .iter()
-                    .filter(|(k, _)| k.starts_with(&prefix))
-                    .map(|(k, v)| (k.clone(), v.clone()))
-                    .collect();
-                slots.sort_by(|a, b| a.0.cmp(&b.0));
-                for (key, value) in slots {
-                    if let Some(slot_hex) = key.strip_prefix(&prefix)
-                        && let Ok(slot) = U256::from_str_radix(slot_hex, 16)
-                    {
-                        // Left-pad short values with leading zeros to 32 bytes
-                        // (big-endian high bits on the left). Note: normally
-                        // values ARE already 32 bytes because we always store
-                        // them that way in writeback, but defensive for any
-                        // legacy entries or state-imports with short blobs.
-                        let mut val_bytes = [0u8; 32];
-                        let n = value.len().min(32);
-                        val_bytes[32 - n..].copy_from_slice(&value[..n]);
-                        let val = U256::from_be_slice(&val_bytes);
-                        evm_db.insert_storage(target_addr, slot, val);
-                    }
-                }
-            }
-        }
-
-        let evm_tx = TxEnv::builder()
-            .caller(from_addr)
-            .kind(tx_kind)
-            .data(alloy_primitives::Bytes::from(calldata))
-            .gas_limit(gas_limit)
-            .gas_price(INITIAL_BASE_FEE as u128)
-            // EVM CREATE/CALL nonce: revm checks `tx.nonce == state.nonce`
-            // and bumps state.nonce internally. Native pass no longer
-            // pre-bumps nonce for EVM txs (charge_fee_only above), so
-            // `sender_nonce` here equals what the EVM tx claimed in its
-            // RLP. Pre-fix this used `sender_nonce.saturating_sub(1)` to
-            // compensate for native double-bump; that was a band-aid that
-            // broke for fresh-wallet CREATE (nonce 0 saturating to 0,
-            // state already at 1 → NonceTooLow). See
-            // `audits/evm-create-nonce-bug-2026-04-27.md`.
-            .nonce(sender_nonce)
-            .chain_id(Some(tx.chain_id))
-            .build()
-            .unwrap_or_default();
-
-        match execute_tx_with_state(evm_db, evm_tx, INITIAL_BASE_FEE, self.chain_id) {
-            Ok((receipt, state)) => {
-                tracing::info!(
-                    "EVM tx {}: success={} gas_used={} contract={:?}",
-                    &tx.txid[..16.min(tx.txid.len())],
-                    receipt.success,
-                    receipt.gas_used,
-                    receipt
-                        .contract_address
-                        .map(|a| format!("0x{}", hex::encode(a.as_slice()))),
-                );
-                if !receipt.success {
-                    // A2: reverted EVM tx — record so eth_getTransactionReceipt
-                    // returns status=0x0 instead of the default 0x1.
-                    self.accounts.mark_evm_tx_failed(&tx.txid);
-                }
-                // Sprint 2: persist every log emitted by this tx. Key is
-                // (height, tx_index, log_index) BE-packed so range scans
-                // return logs in canonical Ethereum order.
-                if let Some(storage) = self.mdbx_storage.as_ref() {
-                    use sentrix_evm::{StoredLog, log_key};
-                    let mut block_hash_bytes = [0u8; 32];
-                    if let Ok(decoded) = hex::decode(block_hash_hex.trim_start_matches("0x")) {
-                        let n = decoded.len().min(32);
-                        block_hash_bytes[..n].copy_from_slice(&decoded[..n]);
-                    }
-                    let mut tx_hash_bytes = [0u8; 32];
-                    if let Ok(decoded) = hex::decode(tx.txid.trim_start_matches("0x")) {
-                        let n = decoded.len().min(32);
-                        tx_hash_bytes[..n].copy_from_slice(&decoded[..n]);
-                    }
-                    for (log_idx, log) in receipt.logs.iter().enumerate() {
-                        let stored = StoredLog::from_revm(
-                            log,
-                            block_height,
-                            block_hash_bytes,
-                            tx_hash_bytes,
-                            tx_index,
-                            log_idx as u32,
-                        );
-                        let key = log_key(block_height, tx_index, log_idx as u32);
-                        let _ =
-                            storage.put_bincode(sentrix_storage::tables::TABLE_LOGS, &key, &stored);
-
-                        // Notify WebSocket subscribers — eth_subscribe(logs).
-                        // Convert StoredLog to the trait-friendly LogData
-                        // shape (sentrix-primitives doesn't depend on
-                        // sentrix-evm so the trait can't take StoredLog
-                        // directly). Non-blocking, infallible.
-                        if let Some(emitter) = &self.event_emitter {
-                            let log_data = sentrix_primitives::events::LogData {
-                                block_height,
-                                block_hash: block_hash_hex.to_string(),
-                                tx_hash: tx.txid.clone(),
-                                tx_index,
-                                log_index: log_idx as u32,
-                                address: stored.address,
-                                topics: stored.topics.clone(),
-                                data: stored.data.clone(),
-                            };
-                            emitter.emit_log(&log_data);
-                        }
-                    }
-                }
-                // Store contract RUNTIME code (not init code) if CREATE succeeded.
-                // receipt.output contains the runtime bytecode returned by the constructor.
-                //
-                // P1 (EIP-170): cap deployed runtime bytecode at 24_576
-                // bytes. Without this an attacker could ship a contract
-                // whose runtime code expands state and trie pages
-                // disproportionately per gas unit, amplifying the cost
-                // of every subsequent block that touches the account.
-                // We reject the CREATE by dropping the stored code and
-                // marking the tx failed; the sender still paid gas but
-                // no contract is registered.
-                const EIP170_MAX_CODE_SIZE: usize = 24_576;
-                if let Some(contract_addr) = receipt.contract_address
-                    && !receipt.output.is_empty()
-                {
-                    if receipt.output.len() > EIP170_MAX_CODE_SIZE {
-                        tracing::warn!(
-                            "P1/EIP-170: contract {} runtime bytecode {} B > {} B limit; \
-                             rejecting deploy (tx {})",
-                            format!("0x{}", hex::encode(contract_addr.as_slice())),
-                            receipt.output.len(),
-                            EIP170_MAX_CODE_SIZE,
-                            &tx.txid[..16.min(tx.txid.len())]
-                        );
-                        self.accounts.mark_evm_tx_failed(&tx.txid);
-                    } else {
-                        let addr_str = format!("0x{}", hex::encode(contract_addr.as_slice()));
-                        use sha3::{Digest as _, Keccak256};
-                        let code_hash: [u8; 32] = Keccak256::digest(&receipt.output).into();
-                        let code_hash_hex = hex::encode(code_hash);
-                        self.accounts
-                            .store_contract_code(&code_hash_hex, receipt.output.clone());
-                        self.accounts.set_contract(&addr_str, code_hash);
-                    }
-                }
-
-                // ── EVM state writeback (closes Bug #1) ─────────────
-                // On success, commit every touched account's balance,
-                // nonce, storage slots, and (for CREATE) bytecode back
-                // to AccountDB so the next tx / next block sees them.
-                // Skip on revert — EVM spec requires state changes to
-                // be discarded on revert; the gas fee was already
-                // debited by the native Pass-1 .transfer() call above,
-                // so there's nothing else to do.
-                //
-                // The EIP-170 guard above may flip the tx to failed
-                // even for a successful revm receipt — in that case
-                // we skip the writeback too. Check mark AFTER the
-                // EIP-170 block ran so "was marked failed" is
-                // authoritative.
-                if receipt.success && !self.accounts.is_evm_tx_failed(&tx.txid) {
-                    commit_state_to_account_db(&state, &mut self.accounts)?;
-                }
-            }
-            Err(e) => {
-                tracing::warn!("EVM tx {} failed: {}", &tx.txid[..16.min(tx.txid.len())], e);
-                // A2: hard execution error — also mark as failed so the
-                // tx receipt reports status=0x0.
-                self.accounts.mark_evm_tx_failed(&tx.txid);
-            }
-        }
-        Ok(())
-    }
 }
 
 // ── Tests ─────────────────────────────────────────────────
